LPS-165285 Inline DefaultFacetProcessor, it serves as a fallback and doesn't need to be a component
